Move file stuff back to plFile for gcc.
