+++ Sources/Plasma/Apps/plClient/plClient.cpp
@@ -58,7 +58,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "hsTimer.h"
 #include "plPipeline/hsG3DDeviceSelector.h"
 #include "plFile/plEncryptedStream.h"
-#include "plFileUtils.h"
+#include "plFile/plFileUtils.h"
 #include "plInputCore/plInputManager.h"
 #include "plInputCore/plInputInterfaceMgr.h"
 #include "plInputCore/plInputDevice.h"
@@ -84,7 +84,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "pfConsole/pfConsoleDirSrc.h"
 #include "plScene/plPageTreeMgr.h"
 #include "plScene/plVisMgr.h"
-#include "hsFiles.h"
+#include "plFile/hsFiles.h"
 
 #include "pfKI/pfKI.h"
