+++ Sources/Plasma/Apps/plPageOptimizer/pfAllCreatables.cpp
@@ -39,16 +39,16 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
       Mead, WA   99021
 
 *==LICENSE==*/
-#include "../pfCharacter/pfCharacterCreatable.h"
-#include "../pfCamera/pfCameraCreatable.h"
-#include "../pfAnimation/pfAnimationCreatable.h"
-#include "../pfConditional/plConditionalObjectCreatable.h"
-//#include "../pfConsole/pfConsoleCreatable.h"
-#include "../pfSurface/pfSurfaceCreatable.h"
-#include "../pfMessage/pfMessageCreatable.h"
-#include "../pfAudio/pfAudioCreatable.h"
-#include "../pfPython/pfPythonCreatable.h"
-#include "../pfGameGUIMgr/pfGameGUIMgrCreatable.h"
-#include "../pfCCR/plCCRCreatable.h"
-#include "../pfJournalBook/pfJournalBookCreatable.h"
-#include "../pfGameMgr/pfGameMgrCreatables.h"
+#include "pfCharacter/pfCharacterCreatable.h"
+#include "pfCamera/pfCameraCreatable.h"
+#include "pfAnimation/pfAnimationCreatable.h"
+#include "pfConditional/plConditionalObjectCreatable.h"
+//#include "pfConsole/pfConsoleCreatable.h"
+#include "pfSurface/pfSurfaceCreatable.h"
+#include "pfMessage/pfMessageCreatable.h"
+#include "pfAudio/pfAudioCreatable.h"
+#include "pfPython/pfPythonCreatable.h"
+#include "pfGameGUIMgr/pfGameGUIMgrCreatable.h"
+#include "pfCCR/plCCRCreatable.h"
+#include "pfJournalBook/pfJournalBookCreatable.h"
+#include "pfGameMgr/pfGameMgrCreatables.h"
