+++ Sources/Plasma/Apps/plPageOptimizer/plPageOptimizer.h
@@ -42,8 +42,8 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #ifndef plPageOptimizer_h_inc
 #define plPageOptimizer_h_inc
 
-#include "../pnKeyedObject/plKey.h"
-#include "../pnKeyedObject/plUoid.h"
+#include "pnKeyedObject/plKey.h"
+#include "pnKeyedObject/plUoid.h"
 #include <vector>
 #include <set>
