+++ Sources/Plasma/CoreLib/CMakeLists.txt
@@ -1,4 +1,3 @@
-include_directories(../PubUtilLib/plUnifiedTime)    # EVIL:  For plFileUtils.cpp
 add_definitions(-D_LIB)
 
 if(NOT WCHAR_BYTES)
@@ -20,7 +19,6 @@ set(CoreLib_SOURCES
     hsCritSect.cpp
     hsExceptionStack.cpp
     hsFastMath.cpp
-    hsFiles.cpp
     hsGeometry3.cpp
     hsMatrix33.cpp
     hsMatrix44.cpp
@@ -36,7 +34,6 @@ set(CoreLib_SOURCES
     hsThread.cpp
     hsWide.cpp
     pcSmallRect.cpp
-    plFileUtils.cpp
     plGeneric.cpp
     plLoadMask.cpp
     plString.cpp
@@ -46,13 +43,11 @@ set(CoreLib_SOURCES
 if(WIN32 AND NOT CYGWIN)
     set(CoreLib_SOURCES ${CoreLib_SOURCES}
         hsThread_Win.cpp
-        hsFiles_Win.cpp
     )
 endif(WIN32 AND NOT CYGWIN)
 if(UNIX)
     set(CoreLib_SOURCES ${CoreLib_SOURCES}
         hsThread_Unix.cpp
-        hsFiles_Unix.cpp
     )
 endif(UNIX)
 
@@ -66,7 +61,6 @@ set(CoreLib_HEADERS
     hsCritSect.h
     hsExceptions.h
     hsFastMath.h
-    hsFiles.h
     hsGeometry3.h
     hsHashTable.h
     hsMatrix44.h
@@ -85,7 +79,6 @@ set(CoreLib_HEADERS
     hsWide.h
     hsWindows.h
     pcSmallRect.h
-    plFileUtils.h
     plGeneric.h
     plLoadMask.h
     plQuality.h
