+++ Sources/Plasma/FeatureLib/pfConsole/pfConsoleCommands.cpp
@@ -155,7 +155,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "plUnifiedTime/plUnifiedTime.h"
 //end for agedefn test
 
-#include "hsFiles.h"
+#include "plFile/hsFiles.h"
 #include "pnSceneObject/plAudioInterface.h"
 
 #include "plStatusLog/plStatusLog.h"
