+++ Sources/Plasma/FeatureLib/pfConsole/pfConsoleCommandsNet.cpp
@@ -84,7 +84,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "plUnifiedTime/plUnifiedTime.h"
 //end for agedefn test
 
-#include "hsFiles.h"
+#include "plFile/hsFiles.h"
 
 #include "plStatusLog/plStatusLog.h"
