+++ Sources/Plasma/PubUtilLib/plAgeDescription/plAgeManifest.cpp
@@ -51,8 +51,8 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "plAgeManifest.h"
 
 
-#include "hsFiles.h"
-#include "plFileUtils.h"
+#include "plFile/hsFiles.h"
+#include "plFile/plFileUtils.h"
 #include "plFile/plInitFileReader.h"
 #include "hsStringTokenizer.h"
