+++ Sources/Plasma/PubUtilLib/plAgeLoader/plResPatcher.cpp
@@ -46,7 +46,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "plAgeLoader/plAgeLoader.h"
 #include "plCompression/plZlibStream.h"
 #include "pnEncryption/plChecksum.h"
-#include "plFileUtils.h"
+#include "plFile/plFileUtils.h"
 #include "plMessage/plResPatcherMsg.h"
 #include "pnNetBase/pnNbError.h"
 #include "plNetGameLib/plNetGameLib.h"
