+++ Sources/Plasma/PubUtilLib/plAudioCore/plSoundBuffer.cpp
@@ -49,8 +49,8 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "plgDispatch.h"
 #include "hsResMgr.h"
 #include "pnMessage/plRefMsg.h"
-#include "plFileUtils.h"
-#include "hsFiles.h"
+#include "plFile/plFileUtils.h"
+#include "plFile/hsFiles.h"
 #include "plUnifiedTime/plUnifiedTime.h"
 #include "plStatusLog/plStatusLog.h"
 #include "hsTimer.h"
