+++ Sources/Plasma/PubUtilLib/plFile/CMakeLists.txt
@@ -5,16 +5,28 @@ include_directories("../../PubUtilLib")
 include_directories(${ZLIB_INCLUDE_DIR})
 
 set(plFile_SOURCES
+    hsFiles.cpp
     plBrowseFolder.cpp
     plEncryptedStream.cpp
+    plFileUtils.cpp
     plInitFileReader.cpp
     plSecureStream.cpp
     plStreamSource.cpp
 )
 
+if(WIN32 AND NOT CYGWIN)
+    set(plFile_SOURCES ${plFile_SOURCES} hsFiles_Win.cpp)
+endif(WIN32 AND NOT CYGWIN)
+
+if(UNIX)
+    set(plFile_SOURCES ${plFile_SOURCES} hsFiles_Unix.cpp)
+endif(UNIX)
+
 set(plFile_HEADERS
+    hsFiles.h
     plBrowseFolder.h
     plEncryptedStream.h
+    plFileUtils.h
     plInitFileReader.h
     plSecureStream.h
     plStreamSource.h
