+++ Sources/Plasma/PubUtilLib/plFile/hsFiles.h
