+++ Sources/Plasma/PubUtilLib/plFile/plFileUtils.cpp
@@ -58,7 +58,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "hsStringTokenizer.h"
 
 
-#include "plUnifiedTime.h"
+#include "plUnifiedTime/plUnifiedTime.h"
 
 #include <time.h>
 #include <sys/types.h>
