+++ Sources/Plasma/PubUtilLib/plNetClient/plNetLinkingMgr.cpp
@@ -63,7 +63,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "plNetMessage/plNetMessage.h"
 #include "plAvatar/plAvatarMgr.h"
 #include "plAvatar/plArmatureMod.h"
-#include "hsFiles.h"
+#include "plFile/hsFiles.h"
 
 
 /*****************************************************************************
