+++ Sources/Plasma/PubUtilLib/plResMgr/plRegistryNode.cpp
@@ -46,7 +46,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "pnKeyedObject/plKeyImp.h"
 #include "plStatusLog/plStatusLog.h"
 #include "pnFactory/plFactory.h"
-#include "plFileUtils.h"
+#include "plFile/plFileUtils.h"
 #include "hsStlUtils.h"
 
 #include "plVersion.h"
