+++ Sources/Plasma/PubUtilLib/plResMgr/plResManager.cpp
@@ -60,8 +60,8 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "pnMessage/plObjRefMsg.h"
 #include "plMessage/plAgeLoadedMsg.h"
 #include "pnMessage/plClientMsg.h"
-#include "hsFiles.h"
-#include "plFileUtils.h"
+#include "plFile/hsFiles.h"
+#include "plFile/plFileUtils.h"
 #include "pnFactory/plCreator.h"
 #include "pnNetCommon/plSynchedObject.h"
 #include "pnNetCommon/plNetApp.h"
