+++ Sources/Plasma/PubUtilLib/plStatGather/plProfileManagerFull.cpp
@@ -50,7 +50,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "hsStream.h"
 #include "pnUtils/pnUtils.h"
 #include "plUnifiedTime/plUnifiedTime.h"
-#include "plFileUtils.h"
+#include "plFile/plFileUtils.h"
 
 plProfileManagerFull::plProfileManagerFull() :
     fVars(plProfileManager::Instance().fVars),
