+++ Sources/Tools/MaxComponent/plAudioComponents.cpp
@@ -42,7 +42,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 
 #include "HeadSpin.h"
 #include "plgDispatch.h"
-#include "plFileUtils.h"
+#include "plFile/plFileUtils.h"
 
 #include "plAnimComponent.h"
 #include "plAudioComponents.h"
