+++ Sources/Tools/MaxComponent/plResponderLink.cpp
@@ -41,7 +41,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 *==LICENSE==*/
 
 #include "HeadSpin.h"
-#include "hsFiles.h"
+#include "plFile/hsFiles.h"
 #include "hsResMgr.h"
 
 #include "plComponentBase.h"
