+++ Sources/Tools/MaxExport/SimpleExport.cpp
@@ -45,7 +45,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 
 #include "HeadSpin.h"
 #include "hsExceptionStack.h"
-#include "plFileUtils.h"
+#include "plFile/plFileUtils.h"
 #include "hsStream.h"
 
 #include <bitmap.h>
