+++ Sources/Tools/MaxMain/Pch.h
@@ -61,7 +61,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "hsColorRGBA.h"
 #include "plgDispatch.h"
 #include "hsFastMath.h"
-#include "hsFiles.h"
+#include "plFile/hsFiles.h"
 #include "hsGeometry3.h"
 #include "pnKeyedObject/plKey.h"
 #include "plLoadMask.h"
