+++ Sources/Tools/MaxMain/main.cpp
@@ -41,7 +41,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 *==LICENSE==*/
 
 #include "HeadSpin.h"
-#include "hsFiles.h"
+#include "plFile/hsFiles.h"
 #include "hsTemplates.h"
 
 #include "MaxComponent/plComponentMgr.h"
