+++ Sources/Tools/MaxMain/plAgeDescInterface.cpp
@@ -40,7 +40,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 
 *==LICENSE==*/
 #include "HeadSpin.h"
-#include "hsFiles.h"
+#include "plFile/hsFiles.h"
 #include "hsStream.h"
 #include "hsTemplates.h"
 #include "hsWindows.h"
